+++ source/current/sha/database/migrations/20251008160000_add_restore_tables.down.sql
@@ -9,3 +9,4 @@ DROP TABLE IF EXISTS restore_mounts;
 
 COMMIT;
 
+