+++ source/current/sha/database/migrations/20251009120000_create_protection_flows.down.sql
@@ -7,3 +7,4 @@ DROP TABLE IF EXISTS protection_flow_executions;
 
 -- Drop main flows table
 DROP TABLE IF EXISTS protection_flows;
+