+++ source/current/sha/database/migrations/20251009160000_add_partition_metadata.down.sql
@@ -4,3 +4,4 @@
 
 ALTER TABLE restore_mounts
 DROP COLUMN partition_metadata;
+